Conflicts:
	Controleur/c_admin.php
	Controleur/c_dashBoard.php
	DAO/JeuDAO.php
	Vue/v_adminliste.php
	Vue/v_header.php

+++ Vue/v_adminliste.php
@@ -85,7 +85,7 @@ if (isset($resultat)) {
                     echo "<a class=\"btn btn-xs btn-success\" href=\"index.php?uc=admin&action=debanUser&id=".$unitem->getIdUser()."\">De-ban l'utilisateur</a>";
                             
                     }?></td>
-                    <td class="mini"><?php echo "<a class=\"btn btn-xs btn-danger\" href=\"index.php?uc=admin&action=deleteUser&id=".$unitem->getIdUser()."\">supprimer utilisateur</a>" ?></td>
+                    <td class="mini"><?php echo "<a class=\"btn btn-xs btn-danger confirm\" href=\"index.php?uc=admin&action=deleteUser&id=".$unitem->getIdUser()."\">supprimer utilisateur</a>" ?></td>
 
     <?php } else if ($titre == "jeux") { ?>
                 <td><?php echo $unitem->getIdJeu(); ?></td>
@@ -95,7 +95,7 @@ if (isset($resultat)) {
                 <td><?php
                     $lesCat = $unitem->getLesCategories();
                     foreach ($lesCat as $unecategorie) {
-                        echo substr($unecategorie, (4)) . "<br>";
+                        echo substr($unecategorie, (7)) . "<br>";
                     }
                     ?></td>
                 <td><?php echo $unitem->getNbJoueurs()->getNbJoueursMin() . " à " . $unitem->getNbJoueurs()->getNbJoueursMax(); ?></td>
@@ -104,14 +104,14 @@ if (isset($resultat)) {
                 <td><?php echo $unitem->getNote(); ?></td>
                 <td><?php echo $unitem->getDateAjout(); ?></td>
                 <td><?php if ($unitem->getImage()) { ?><img width="150" height="100" src="Vue/img/jeu/<?php echo $unitem->getImage(); ?>" /><?php } ?></td>
-                <td><?php echo "<a class=\"btn btn-xs btn-danger\" href=\"index.php?uc=admin&action=deleteJeu&id=".$unitem->getIdJeu()."\">supprimer jeu</a>" ?></td>
+                <td><?php echo "<a class=\"btn btn-xs btn-danger confirm\" href=\"index.php?uc=admin&action=deleteJeu&id=".$unitem->getIdJeu()."\">supprimer jeu</a>" ?></td>
     <?php } else if ($titre == "évènements") { ?>
                 <td><?php echo $unitem->getIdEvenement(); ?></td>
                 <td><?php echo $unitem->getTitre(); ?></td>
                 <td><?php echo htmlentities($unitem->getEvenement()); ?></td>
                 <td><?php if ($unitem->getLienImage()) { ?><img width="200" height="100" src="<?php echo $unitem->getLienImage(); ?>" /><?php } ?></td>
                 <td><?php echo $unitem->getDateAjout(); ?></td>
-                <td><?php echo "<a class=\"btn btn-xs btn-danger\" href=\"index.php?uc=admin&action=deleteEvenement&id=".$unitem->getIdEvenement()."\">Supprimer evenement</a>" ?></td>
+                <td><?php echo "<a class=\"btn btn-xs btn-danger confirm\" href=\"index.php?uc=admin&action=deleteEvenement&id=".$unitem->getIdEvenement()."\">Supprimer evenement</a>" ?></td>
                 <td><?php echo "<a class=\"btn btn-xs btn-danger\" href=\"index.php?uc=admin&action=demModifEvenement&id=".$unitem->getIdEvenement()."\">Modifier evenement</a>" ?></td>
                     <?php
             }
@@ -137,7 +137,7 @@ if (isset($resultat)) {
                     }
                     ?></td>
                 <td class="mini"><?php echo "<a class=\"btn btn-xs btn-success\" href=\"index.php?uc=admin&action=debanUser&id=".$unitem->getIdUser()."\">De-ban l'utilisateur</a>" ?>
-                <td class="mini">    <?php echo "<a class=\"btn btn-xs btn-danger\" href=\"index.php?uc=admin&action=deleteUser&id=".$unitem->getIdUser()."\">supprimer utilisateur</a>" ?>
+                <td class="mini">    <?php echo "<a class=\"btn btn-xs btn-danger confirm\" href=\"index.php?uc=admin&action=deleteUser&id=".$unitem->getIdUser()."\">supprimer utilisateur</a>" ?>
                 </td>
                 </td>
     <?php
@@ -146,5 +146,7 @@ if (isset($resultat)) {
 ?>
     </tr>
 </table>
+
+
 
 <?php include('v_footer.php'); ?>
